+++ packages/core/src/__tests__/angular.test.ts
@@ -1,5 +1,5 @@
 import { componentToAngular } from '../generators/angular';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Angular', () => {
   runTestsForTarget({ options: {}, target: 'angular', generator: componentToAngular });
