+++ packages/core/src/__tests__/html.test.ts
@@ -1,5 +1,5 @@
 import { componentToHtml } from '../generators/html';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Html', () => {
   runTestsForTarget({ options: {}, target: 'html', generator: componentToHtml });
