+++ packages/core/src/__tests__/lit.test.ts
@@ -1,5 +1,5 @@
 import { componentToLit } from '../generators/lit';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Lit', () => {
   runTestsForTarget({ options: {}, target: 'lit', generator: componentToLit });
