+++ packages/core/src/__tests__/parse-svelte.test.ts
@@ -1,4 +1,4 @@
-import { runTestsForSvelteSyntax } from './shared';
+import { runTestsForSvelteSyntax } from './test-generator';
 
 describe('Parse JSX', () => {
   runTestsForSvelteSyntax();
