+++ packages/core/src/__tests__/react-native.test.ts
@@ -1,5 +1,5 @@
 import { componentToReactNative } from '../generators/react-native';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('React Native', () => {
   runTestsForTarget({ options: {}, target: 'reactNative', generator: componentToReactNative });
