+++ packages/core/src/__tests__/react.test.ts
@@ -1,6 +1,6 @@
 import { parseJsx } from '..';
 import { componentToReact } from '../generators/react';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 import stamped from './data/blocks/stamped-io.raw.tsx?raw';
