+++ packages/core/src/__tests__/rsc.test.ts
@@ -1,5 +1,5 @@
 import { componentToRsc } from '../generators/rsc';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('RSC', () => {
   runTestsForTarget({ options: {}, target: 'rsc', generator: componentToRsc });
