+++ packages/core/src/__tests__/svelte.test.ts
@@ -1,5 +1,5 @@
 import { componentToSvelte } from '../generators/svelte';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Svelte', () => {
   runTestsForTarget({ target: 'svelte', generator: componentToSvelte, options: {} });
