+++ packages/core/src/__tests__/taro.test.ts
@@ -1,5 +1,5 @@
 import { componentToTaro } from '../generators/taro';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Taro', () => {
   runTestsForTarget({ options: {}, target: 'taro', generator: componentToTaro });
