+++ packages/core/src/__tests__/vue-composition.test.ts
@@ -1,5 +1,5 @@
 import { componentToVue3 } from '../generators/vue';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Vue', () => {
   runTestsForTarget({ target: 'vue', generator: componentToVue3, options: { api: 'composition' } });
