+++ packages/core/src/__tests__/vue.test.ts
@@ -1,5 +1,5 @@
 import { componentToVue2 } from '../generators/vue';
-import { runTestsForTarget } from './shared';
+import { runTestsForTarget } from './test-generator';
 
 describe('Vue', () => {
   runTestsForTarget({ options: { api: 'options' }, target: 'vue', generator: componentToVue2 });
